refactor(l1): satisfy Codacy complexity gate
